+++ src/cart/index.html
@@ -27,6 +27,6 @@
       </section>
     </main>
 
-    <footer>&copy;NOT a real business</footer>
+    <footer id="main-footer">&copy;NOT a real business</footer>
   </body>
 </html>
