+++ src/js/ProductList.mjs
@@ -27,14 +27,6 @@ export default class ProductList {
 
     async init() {
         const list = await this.dataSource.getData();
-        // this.renderList(list);
         renderListWithTemplate(productCardTemplate,this.listElement,list);
-
     }
-
-    // renderList(list) {
-    //     // this.listElement.innerHTML = list.map(productCardTemplate).join("")//my original way
-    //     const htmlStrings = list.map(productCardTemplate);
-    //     this.listElement.insertAdjacentHTML('afterbegin', htmlStrings.join(''));
-    // }
 }
