+++ src/js/cart.js
@@ -1,28 +1,33 @@
-import { getLocalStorage, loadHeaderFooter } from "./utils.mjs";
+import {loadHeaderFooter, } from "./utils.mjs";
+import ShoppingCart from "./ShoppingCart.mjs";
 
-function renderCartContents() {
-  const cartItems = getLocalStorage("so-cart");
-  const htmlItems = cartItems.map((item) => cartItemTemplate(item));
-  document.querySelector(".product-list").innerHTML = htmlItems.join("");
-}
+const cartItems = new ShoppingCart( document.querySelector(".product-list"))
 
-function cartItemTemplate(item) {
-  const newItem = `<li class="cart-card divider">
-  <a href="#" class="cart-card__image">
-    <img
-      src="${item.Image}"
-      alt="${item.Name}"
-    />
-  </a>
-  <a href="#">
-    <h2 class="card__name">${item.Name}</h2>
-  </a>
-  <p class="cart-card__color">${item.Colors[0].ColorName}</p>
-  <p class="cart-card__quantity">qty: 1</p>
-  <p class="cart-card__price">$${item.FinalPrice}</p>
-</li>`;
-
-  return newItem;
-}
+cartItems.init();
 loadHeaderFooter();
-renderCartContents();
+// function renderCartContents() {
+//   const cartItems = getLocalStorage("so-cart");
+//   const htmlItems = cartItems.map((item) => cartItemTemplate(item));
+//   document.querySelector(".product-list").innerHTML = htmlItems.join("");
+// }
+
+// function cartItemTemplate(item) {
+//   const newItem = `<li class="cart-card divider">
+//   <a href="#" class="cart-card__image">
+//     <img
+//       src="${item.Image}"
+//       alt="${item.Name}"
+//     />
+//   </a>
+//   <a href="#">
+//     <h2 class="card__name">${item.Name}</h2>
+//   </a>
+//   <p class="cart-card__color">${item.Colors[0].ColorName}</p>
+//   <p class="cart-card__quantity">qty: 1</p>
+//   <p class="cart-card__price">$${item.FinalPrice}</p>
+// </li>`;
+
+//   return newItem;
+// }
+
+// renderCartContents();
